requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
beautifulsoup4==4.12.2
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
docopt==0.6.2
Faker==20.1.0
googlesearch-python==1.2.3
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.10
nltk==3.8.1
num2words==0.5.13
numpy==1.26.2
packaging==23.2
pandas==2.1.3
preshed==3.0.9
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
textblob==0.17.1
thinc==8.2.1
threadpoolctl==3.2.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
wikipedia==1.4.0
